feat: add conditional compilation for attohttpc and rayon.
Implemented conditional compilation for 'attohttpc' and 'rayon' dependencies. These libraries are now excluded from the build when 'default_feature' is set to 'false'.
