+++ src/lib.rs
@@ -55,6 +55,8 @@ mod map;
 /// Tools for drawing features onto the map.
 pub mod tools;
 
+pub mod fetchers;
+
 pub use bounds::Bounds;
 pub use error::Error;
 pub use map::{StaticMap, StaticMapBuilder};
